refactor(caption): remove CaptionDatasetAdapter base class, make adapters self-contained
- Delete caption/base.py (598-line base class), replace with caption/metrics.py
  containing only compute_caption_scores + tokenizer helpers (~100 lines)
- Rewrite msr_vtt_adapter.py as self-contained MVBench-style adapter
- Rewrite msvd_adapter.py as self-contained adapter, dataset_id → evalscope/MSVD
- Rewrite vqav2_adapter.py as standard DocVQA-style adapter using DefaultDataAdapter pipeline
- Revert hub.py subset=None change (unnecessary for MSR-VTT)
- Add DashScope fps extension comment in openai.py
- Split tests into test_caption.py + test_vqav2.py, test public interfaces
